Use a more reasonable logic for finding the next line
I was going to archive this repo but then I looked at the code and
found it too naive :-) This patch changes the logic to scan on
request instead of building a list of long lines just to immediately
discard them.

Now I can archive this with a peaceful mind.